Restructure API folder
